refactor(www): panel-lab per-section files, tabs, drop v1
One section is now one file in sections/, owning its body, its options
and its defaults slice. The defaults slice is the state slice: section
reset spreads it back, so the separate *_KEYS arrays are gone. state.ts
is the only shared wiring — it spreads the slices into DEFAULTS and
orders CHAPTERS. Shape owns radius resolution; Space, Focus, Buttons and
Inputs read roleRadiusPx from it.

v1 and the drafts are deleted along with versions.tsx, gallery.tsx,
version-page.tsx and the $version/draft/controls routes. /internal/panel-lab
becomes one tabbed page: Panel (the former v2) and Primitives (the row
catalog, previously /internal/panel-lab/controls).
